refactor(clawdstrike-policy-event): extract conversion from edr/mod.rs
Move the final free-function cluster into edr/conversion.rs: content-hash
helpers (canonical_graph_content_hash, endpoint_*_content_hash,
observation_age_seconds), the PolicyEvent->Endpoint conversion
(endpoint_event_from_policy_event, metadata_as_btree, promote_* identity
context), process_from_metadata, and the metadata field-extractor family
(u32_field, string_field, *_id_field, tool_*_field, credential_kind_*,
event_target_*, process_*_hash_field, string_field_nested).

These are reached cross-module by event.rs, actor.rs, causal/recorder.rs,
flight_recorder/mod.rs, receipt/{mod,evidence}, and detection/supply_chain
via super:: paths, so the 21 cross-module items are re-exported pub(crate)
at the edr root (private fns widened to pub(crate) as needed; the 13
internal helpers stay private). conversion reaches finding_builders
(credential_kind_from_secret) and the edr types via super::. The three
flight-recorder schema-version consts remain in mod.rs.

mod.rs is now a ~98-line re-export hub. Pure relocation; bodies unchanged.
